Typed select items and restored values reloading
Because we want to have mostly automated GraphQL support, that means
we need a structured and predictable selection value. So having a
predictable, mostly scalar, selection value, means that we can never
return abitrary object to the client. So there is no reason to allow
arbitrary objects as configuration, because they will never be used.

Instead we defined a fixed interface that should cover most cases, and
it is up to the user to map whatever his model is before configuring
natural-search.